backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart>=0.0.9
sqlalchemy==2.0.36
psycopg[binary]>=3.1
python-dotenv==1.0.1
pyjwt>=2.8
bcrypt>=4.0
email-validator>=2.0
alembic>=1.14
